Extract Glue client provider in Hive
